fix: prevent watchOS Int overflow in session_expiry parsing
`10_000_000_000` overflows a 32-bit `Int` literal on watchOS (armv7k /
arm64_32), so 2.24.0 failed to compile when archiving an app that embeds a
watchOS companion target.

Bound the value in the `Double` domain before truncating, and return `Int64`
instead of `Int`. A valid `session_expiry` ceiling can exceed `Int32.max`
(any timestamp after 2038-01-19), which would either overflow the literal at
compile time or trap at runtime in `Int(rawValue)` on watchOS's 32-bit `Int`.
`Int64` sidesteps both regardless of platform word size. The pinned-ceiling
reader is widened to `Int64` to match. These are package-internal APIs, so the
signature change is not a public break. Behavior is unchanged on all platforms.

Fixes #1240

+++ Auth0/CredentialsManager.swift
@@ -1061,10 +1061,10 @@ public struct CredentialsManager: Sendable {
     }
 
     /// Returns the `session_expiry` ceiling (Unix seconds) pinned at login, or `nil` if nothing is stored.
-    private func pinnedSessionExpiry() -> Int? {
+    private func pinnedSessionExpiry() -> Int64? {
         guard let data = self.storage.getEntry(forKey: self.sessionExpiryKey),
               let string = String(data: data, encoding: .utf8),
-              let value = Int(string),
+              let value = Int64(string),
               value > 0 else { return nil }
         return value
     }
